Fix critical crash and refactor to modern Android architecture
This commit addresses several critical issues and modernizes the app's architecture for better stability, maintainability, and performance.

Key changes:
- **Fix Critical Crash:** Patched `AudioMeter.kt` to prevent a `NoSuchMethodError` on Android versions below API 31 by using the correct `MediaRecorder` constructor based on SDK version.
- **Implement ViewModel:** Introduced `MainViewModel` to manage all application state and business logic. This resolves state loss on configuration changes (e.g., screen rotation) and separates concerns.
- **Refactor MainActivity:** The `MainActivity` now serves as a lean UI controller, observing state from the `MainViewModel` via `StateFlow`. All logic has been moved into the ViewModel.
- **Improve Robustness:** Eliminated a potential crash on app startup by making the Spinner initialization robust against empty data sources.
- **Refine Resource Management:** Replaced hardcoded Holo colors with modern color definitions in `colors.xml`.
- **Cleanup Dependencies:** Removed unused Jetpack Compose dependencies and plugins from the Gradle configuration, reducing project size and build complexity.

+++ app/src/main/java/nl/jeoffrey/geluidsboetechecker/data/VehicleLimits.kt
@@ -3,14 +3,14 @@ package nl.jeoffrey.geluidsboetechecker.data
 object VehicleLimits {
 
     private val limits = mapOf(
-        "Auto" to (70.0 to 90.0),
-        "Motor" to (75.0 to 95.0),
-        "Brommer" to (72.0 to 92.0)
+        "Auto" to (70 to 90),
+        "Motor" to (75 to 95),
+        "Brommer" to (72 to 92)
     )
 
-    private val defaultLimits = 70.0 to 90.0
+    private val defaultLimits = 70 to 90
 
-    fun getLimitsFor(vehicle: String): Pair<Double, Double> {
+    fun getLimitsFor(vehicle: String): Pair<Int, Int> {
         return limits.getOrDefault(vehicle, defaultLimits)
     }
 }
